util/file: move TeeReadCloser from provisioning
