+++ internal/util/file/tee_read_closer.go
@@ -1,4 +1,4 @@
-package provisioning
+package file
 
 import "io"
