+++ internal/util/file/tee_read_closer_internal_test.go
@@ -1,4 +1,4 @@
-package provisioning
+package file
 
 import (
 	"bytes"
